requirements.txt
mysql-connector
html-similarity
